FIxed some issues that came up whilst having it monitor a tournamnet. Console now also counts down the time to the next check,
Updated the the checks for the top cut stuff as when there isn't a 3rd
place match I was worried this would cause issues.

Matches now get assigned the is_finals check correctly

The ID for attendance_id is now assigned properly.

init_db no longer creates multiples of the same values and values are
added after the creation of the Database